% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_aggregation.R
\name{consensus_aggregation}
\alias{consensus_aggregation}
\title{Create Index of randomly sampling in aggregation}
\usage{
consensus_aggregation(N, M, seed, rep, id_name, bisbmresults, id_label, ...)
}
\arguments{
\item{N}{Aggregation times}

\item{M}{The number of total sbm results}

\item{seed}{A vector of the seed you want to set for each aggregation time}

\item{rep}{A vector of the number of pairs for comparison at each aggregation}

\item{id_name}{name of the variable you want to cluster}

\item{bisbmresults}{The results from run_bisbm}

\item{id_label}{The label of the nodes}
}
\value{
the list of a pair of independent consensus graph with the same aggregation times
}
\description{
Given the number of M affinity matrix from bisbm results, randomly select N1 results to construct the
consensus graph and compare with the consensus graph which constructed by another N2 results
that randomly selected in the rest of all M-N1 results.
}
\details{
\code{consensus_aggregation} returns the list of a pair of independent consensus graph with the same aggregation times
with the same number of aggregation times
}
\examples{
data("bisbmresults")
library(dplyr,quietly=TRUE,warn.conflicts=FALSE)

id_label <- bisbmresults[[1]] \%>\%
         filter(type=="id") \%>\%
         filter(level==0) \%>\%
         dplyr::select(node) \%>\%
         arrange(node)
id_label <- id_label$node

consensus_matrix_list <- consensus_aggregation(5,10,c(1:5),5,"id",bisbmresults,id_label)
}
