% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_spectral.R
\name{consensus_spectral}
\alias{consensus_spectral}
\title{NMI measurement to check the convergence of aggregation times}
\usage{
consensus_spectral(consensus_matrix_list, method)
}
\arguments{
\item{consensus_matrix_list}{the list of a pair of independent consensus graph with the same aggregation times}

\item{method}{The partition method in spectral clustering}
}
\value{
a list of NMI measurement for a pair of independent consensus graph with the same aggregation times
}
\description{
\code{consensus_spec} returns a pair of consensus graphs
with the same number of aggregation times
}
\examples{
consensus_spec <- consensus_spectral(consensus_matrix_list,)
}
