% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_tree.R
\name{data_for_tree}
\alias{data_for_tree}
\title{Generate json format data for binary tree visualization}
\usage{
data_for_tree(comm_member)
}
\arguments{
\item{comm_member}{Long format community membership file from `comm_member` function}
}
\value{
a json file for binary tree
}
\description{
\code{data_for_tree} returns a json file for binary tree
}
\examples{
data("cluster.result")

comm_member <- community_membership(cluster.result$tree.path,
nlabels=sum(sapply(cluster.result$cluster.labels.list,length)))
data_for_tree <- data_for_tree(comm_member)
}
