% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bisbm.R
\name{run_bisbm}
\alias{run_bisbm}
\title{Run graph-tool based bisbm parallelly}
\usage{
run_bisbm(ntimes, ncore, dat, ...)
}
\arguments{
\item{ntimes}{The number of times for running sbm}

\item{ncore}{The number of cores required}

\item{dat}{The data which satisfy the requirements of the format of runnning sbm}
}
\value{
Lists of sbm results, 
        each element of the list is a dataframe of biSBM cluster results
}
\description{
\code{run_bisbm} returns lists of bisbm results
}
\examples{
#run bisbm twp times
data("phenome_dat_long")
library(dplyr,quietly=TRUE)

head(phenome_dat_long)

bisbmresults <- run_bisbm(2,2,phenome_dat_long)
head(bisbmresults[[1]])
}
