% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_aggregation.R
\name{consensus_aggregation}
\alias{consensus_aggregation}
\title{Create Index of randomly sampling in aggregation}
\usage{
consensus_aggregation(N, M, seed, rep, bisbmresults, id_label)
}
\arguments{
\item{N}{Aggregation times}

\item{M}{The number of total sbm results}

\item{seed}{A vector of the seed you want to set for each aggregation time}

\item{rep}{The number of pairs for comparison}

\item{bisbmresults}{The results from run_bisbm}

\item{id_label}{The label of the nodes}
}
\value{
the list of a pair of independent consensus graph with the same aggregation times
}
\description{
Given the number of M affinity matrix from bisbm results, randomly select N results to construct the
consensus graph and compare with the consensus graph which constructed by another N results
that randomly selected in the rest of M-first selected N results.
}
\details{
\code{consensus_aggregation} returns the list of a pair of independent consensus graph with the same aggregation times
with the same number of aggregation times
}
\examples{
consensus_matrix_list <- consensus_aggregation(20,500,c(1:50),50,bisbmresults,id_label)
}
