% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_build.R
\name{consensus_build}
\alias{consensus_build}
\title{Build the consensus graph}
\usage{
consensus_build(id_label, bisbm)
}
\arguments{
\item{id_label}{The lable of nodes}

\item{bisbm}{The list of bisbm results from run_bisbm}
}
\value{
The long format of consensus matrix (node to node connection weight)
}
\description{
\code{consensus_build} returns the long format of consensus matrix (node to node connection weight)
}
